Add Angular frontend and SDK live auth updates

+++ f1-sdk/f1_sdk/Models/wather.py
@@ -6,7 +6,7 @@ from .f1BaseModel import F1BaseModel
 class Weather(F1BaseModel):
     air_temperature: float
     date: str
-    humidity: int
+    humidity: float
     meeting_key: int
     pressure: float
     rainfall: int
